:hammer: [chore] HomeView 추가 #12
