+++ coupleDayProject/Common/Constants.swift
@@ -232,3 +232,74 @@ final class Observable<T> {
 		listener = closure
 	}
 }
+
+/// 기념일 present 할 때 위에 어느정도 띄워주는 custom modal
+final class PresentationController: UIPresentationController {
+	
+	let blurEffectView: UIVisualEffectView!
+	var tapGestureRecognizer: UITapGestureRecognizer = UITapGestureRecognizer()
+	
+	override init(
+		presentedViewController: UIViewController,
+		presenting presentingViewController: UIViewController?
+	) {
+		let blurEffect = UIBlurEffect(style: .dark)
+		blurEffectView = UIVisualEffectView(effect: blurEffect)
+		super.init(
+			presentedViewController: presentedViewController,
+			presenting: presentingViewController
+		)
+		tapGestureRecognizer = UITapGestureRecognizer(
+			target: self,
+			action: #selector(dismissController)
+		)
+		blurEffectView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+		self.blurEffectView.isUserInteractionEnabled = true
+		self.blurEffectView.addGestureRecognizer(tapGestureRecognizer)
+	}
+	
+	override var frameOfPresentedViewInContainerView: CGRect {
+		CGRect(
+			origin: CGPoint(x: 0, y: self.containerView!.frame.height * 0.15),
+			size: CGSize(
+				width: self.containerView!.frame.width,
+				height: self.containerView!.frame.height * 0.85
+			)
+		)
+	}
+	
+	override func presentationTransitionWillBegin() {
+		self.blurEffectView.alpha = 0
+		self.containerView?.addSubview(blurEffectView)
+		self.presentedViewController.transitionCoordinator?.animate(
+			alongsideTransition: { (UIViewControllerTransitionCoordinatorContext) in
+				self.blurEffectView.alpha = 0.7
+			}, completion: { (UIViewControllerTransitionCoordinatorContext) in }
+		)
+	}
+	
+	override func dismissalTransitionWillBegin() {
+		self.presentedViewController.transitionCoordinator?.animate(
+			alongsideTransition: { (UIViewControllerTransitionCoordinatorContext) in
+				self.blurEffectView.alpha = 0
+			}, completion: { (UIViewControllerTransitionCoordinatorContext) in
+				self.blurEffectView.removeFromSuperview()
+			}
+		)
+	}
+	
+	override func containerViewWillLayoutSubviews() {
+		super.containerViewWillLayoutSubviews()
+		presentedView!.roundCorners([.topLeft, .topRight], radius: 22)
+	}
+	
+	override func containerViewDidLayoutSubviews() {
+		super.containerViewDidLayoutSubviews()
+		presentedView?.frame = frameOfPresentedViewInContainerView
+		blurEffectView.frame = containerView!.bounds
+	}
+	
+	@objc func dismissController(){
+		self.presentedViewController.dismiss(animated: true, completion: nil)
+	}
+}
